script/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
bio==0.4.1
biopython==1.78
Bottleneck==1.3.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.17
grpcio==1.32.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.0.1
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
missingno==0.4.2
multimethod==1.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.3
-e git+git@github.com:pandas-profiling/pandas-profiling.git@17523d8e81f5841ebda9f80e1977b7fc26e8afef#egg=pandas_profiling
phik==0.11.2
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
tangled-up-in-unicode==0.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.0
typing-extensions==3.7.4.3
urllib3==1.26.4
visions==0.7.1
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==0.90
zipp==3.4.1
